Rename DiverseSelector to selector
