+++ selector/tests/common.py
